Thesis_MMFF/requirements.txt
torch
torchvision
timm
numpy
pillow
tqdm
scikit-learn
matplotlib
